+++ RELEASENOTES.md
@@ -5,6 +5,7 @@ These are the release notes for the [mediawiki-api](README.md).
 * Introduces RevisionUndoer service
 * Introduces UserCreator service
 * Introduces FileUploader service
+* Introduces ImageRotator service
 * Using mediawiki-api-base 0.6
 * Using mediawiki-datamodel 0.3